+++ src/backend/src/controllers/recruitment.controllers.ts
@@ -96,14 +96,4 @@ export default class RecruitmentController {
       return next(error);
     }
   }
-
-  static async deleteChecklist(req: Request, res: Response, next: NextFunction) {
-    try {
-      const { checklistId } = req.params;
-      await RecruitmentServices.deleteChecklist(req.currentUser, checklistId, req.organization);
-      res.status(200).json({ message: `Successfully deleted checklist with id ${checklistId}` });
-    } catch (error: unknown) {
-      return next(error);
-    }
-  }
 }
