+++ src/backend/src/routes/onboarding.routes.ts
@@ -14,4 +14,6 @@ onboardingRouter.post(
   OnboardingController.createChecklist
 );
 
+onboardingRouter.delete('/checklist/:checklistId/delete', OnboardingController.deleteChecklist);
+
 export default onboardingRouter;
