+++ src/backend/src/routes/recruitment.routes.ts
@@ -50,7 +50,4 @@ recruitmentRouter.post(
 
 recruitmentRouter.delete('/faq/:faqId/delete', RecruitmentController.deleteFaq);
 
-/* Checklist Section */
-recruitmentRouter.delete('/checklist/:checklistId/delete', RecruitmentController.deleteChecklist);
-
 export default recruitmentRouter;
