+++ src/backend/tests/unmocked/recruitment.test.ts
@@ -337,41 +337,4 @@ describe('Recruitment Tests', () => {
       expect(deletedTestFaq?.dateDeleted).not.toBe(null);
     });
   });
-
-  describe('Delete Checklist', () => {
-    it('Fails if user is not admin', async () => {
-      await expect(
-        async () =>
-          await RecruitmentServices.deleteChecklist(await createTestUser(wonderwomanGuest, orgId), 'id', organization)
-      ).rejects.toThrow(new AccessDeniedAdminOnlyException('delete a checklist'));
-    });
-
-    it('Fails if checklistId is not found', async () => {
-      await expect(
-        async () =>
-          await RecruitmentServices.deleteChecklist(await createTestUser(batmanAppAdmin, orgId), 'id1', organization)
-      ).rejects.toThrow(new HttpException(400, 'Checklist with id: id1 not found!'));
-    });
-
-    it('Fails if checklist is already deleted', async () => {
-      const testSuperman = await createTestUser(supermanAdmin, orgId);
-      const testChecklist = await createTestChecklist(testSuperman, orgId);
-      await RecruitmentServices.deleteChecklist(testSuperman, testChecklist.checklistId, organization);
-
-      await expect(
-        async () => await RecruitmentServices.deleteChecklist(testSuperman, testChecklist.checklistId, organization)
-      ).rejects.toThrow(new DeletedException('Checklist', testChecklist.checklistId));
-    });
-
-    it('Succeeds and deletes checklist', async () => {
-      const testSuperman = await createTestUser(supermanAdmin, orgId);
-      const testChecklist1 = await createTestChecklist(testSuperman, orgId);
-      await RecruitmentServices.deleteChecklist(testSuperman, testChecklist1.checklistId, organization);
-
-      const updatedTestChecklist1 = await prisma.checklist.findUnique({
-        where: { checklistId: testChecklist1.checklistId }
-      });
-      expect(updatedTestChecklist1?.dateDeleted).not.toBe(null);
-    });
-  });
 });
